@@ -1,10 +1,11 @@
-import Vue from 'vue'
+import { createApp, h } from 'vue';
 import App from './App.vue'
 import vuetify from './plugins/vuetify';
+import router from './router'
+import store from './store'
 
-Vue.config.productionTip = false
-
-new Vue({
-  vuetify,
-  render: h => h(App),
-}).$mount('#app')
+createApp(App)
+  .use(vuetify)
+  .use(router)
+  .use(store)
+  .mount('#app')
